changed: made all subviews of default template public
changed: made WithBackgroundColor public
changed: the paywall auto selects weekly product on init
changed: paywallstore made products and selectedProduct public
