+++ Sources/AwesomePaywall/Views/WithBackgroundColor.swift
@@ -8,14 +8,14 @@
 import SwiftUI
 
 // Use background color full or as gradient
-struct WithBackgroundColor<Content: View>: View {
+public struct WithBackgroundColor<Content: View>: View {
     let color: Color
     var asGradient: Bool = false
     var gradientStart: UnitPoint = .top
     var gradientStop: UnitPoint = .bottom
     @ViewBuilder let content: () -> Content
 
-    init(color: Color, asGradient: Bool = false, gradientStart: UnitPoint = .top, gradientStop: UnitPoint = .bottom, content: @escaping () -> Content) {
+    public init(color: Color, asGradient: Bool = false, gradientStart: UnitPoint = .top, gradientStop: UnitPoint = .bottom, content: @escaping () -> Content) {
         self.color = color
         self.content = content
         self.asGradient = asGradient
@@ -23,7 +23,7 @@ struct WithBackgroundColor<Content: View>: View {
         self.gradientStop = gradientStop
     }
 
-    var body: some View {
+    public var body: some View {
         if self.asGradient {
             self.content()
                 .background(.linearGradient(self.color.gradient, startPoint: self.gradientStart, endPoint: self.gradientStop))
